refactor: move executable resolution functions to `executable-path-resolver.ts`
- Move `resolveCBFExecutablePath`, `resolveCSExecutablePath`, and `resolvePhpExecutablePath` functions from settings into a new `executable-path-resolver.ts` file for better organisation, and separating executable resolvers from settings.

- Changed the file imports as necessary.
